fix(win): LLP64 format mismatches, and quiet the Windows build's noise
Nine printf sites passed size_t or sizeof() to %ld / %lu.  On LP64 — Linux and
macOS — long and size_t are the same width, so clang has never had anything to
say about them.  Windows is LLP64: long is 32 bits, size_t is 64.  Every one of
those log lines prints a truncated number and leaves the vararg list misaligned
for whatever follows it.  Casting to (long)/(unsigned long) at the call site
fixes it on all three platforms and, unlike switching to %zu, does not depend on
which printf implementation the CRT supplies.

hfdl_ohma.c's %td goes the same way: it is C99, and the Microsoft printf is not
obliged to know it.

Also for warning hygiene, which is the whole reason the Makefile suppresses the
idiomatic-noise categories: WDSP's linux_port.h defines away __declspec,
__cdecl, __stdcall and __forceinline, which gcc predefines on Windows — a
redefinition warning in every translation unit that reaches the header, 248
lines of it.  They are now #undef'd first rather than skipped, so the shim's
empty versions still win: letting gcc's real __declspec stand would turn WDSP's
PORT into __declspec(dllexport), and naming even one export explicitly switches
OFF mingw's auto-export of everything else — the DLL would quietly lose symbols.

-DNOMINMAX stops <windef.h> defining min/max over the ones main.h and
linux_port.h declare.

With HFDL enabled (using the liquid-dsp header, which is platform-independent),
the whole tree — decoders included — compiles for Windows with zero errors.  The
only thing still missing is the liquid-dsp binary itself, which MSYS2 does not
package and which needs a source build.

The 40 warnings that remain are gcc-only and pre-existing (misleading
indentation, maybe-uninitialized, format truncation, implicit fallthrough); they
are not Windows-specific and a Linux gcc build shows the same set.

macOS build clean.

+++ src/decode/hfdl_ohma.c
@@ -334,7 +334,7 @@ gboolean hfdl_ohma_decode(const char *reg, const char *txt, GString *out, int in
     if (guard >= 4) break;
     const char *dup = find_mem(ptr, len, txt, (gsize)(ptr - txt));
     if (dup == NULL) break;
-    log_debug("hfdl ohma: duplicate first fragment, skipping %td octets\n", dup - txt);
+    log_debug("hfdl ohma: duplicate first fragment, skipping %ld octets\n", (long)(dup - txt));
     ptr = dup;
   }
